requirements.txt
altair==5.3.0
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cramjam==2.8.3
cryptography==42.0.5
cycler==0.12.1
dacite==1.8.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
et-xmlfile==1.1.0
Faker==25.3.0
fastapi==0.110.0
fastavro==1.9.4
fonttools==4.51.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.5
httpx==0.26.0
idna==3.7
ImageHash==4.3.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.6
kiwisolver==1.4.5
llvmlite==0.42.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
multimethod==1.11.2
networkx==3.3
numba==0.59.1
numpy==1.26.4
numpyencoder==0.3.0
oauthlib==3.2.2
openpyxl==3.1.5
oracledb==2.1.1
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
patsy==0.5.6
pefile==2024.8.26
phik==0.12.4
pillow==10.3.0
protobuf==4.25.3
psycopg2==2.9.9
pyarrow==15.0.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
pydrill==0.3.4
Pygments==2.18.0
pyhdb==0.3.4
PyHive==0.7.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pymongo==4.6.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyodbc==5.1.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-jose==3.3.0
python-keycloak==4.0.0
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
pywin32-ctypes==0.2.3
PyYAML==6.0.1
referencing==0.35.1
reportlab==4.2.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
starlette==0.36.3
statsmodels==0.14.2
streamlit==1.35.0
tabula-py==2.9.3
tenacity==8.3.0
thrift==0.20.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
typeguard==4.2.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
visions==0.7.6
watchdog==4.0.1
wordcloud==1.9.3
xmltodict==0.13.0
ydata-profiling==4.8.3